Modify test.rs. Fixed tests
